Crate asn1

This library provides tools for decoding/encoding ASN.1 messages to/from their corresponding Rust types.

ASN.1 Type Correspondence

Below are the currently supported ASN.1 types and their corresponding types/constructs in Rust.

ASN.1 Type 	Rust Type

BIT STRING 	BitString

INTEGER* 	i8,i16,i32,u8,u16,u32

NULL 	()

OCTET STRING 	Vec<u8>

SEQUENCE 	struct

SEQUENCE OF 	Vec<T>

CHOICE 	enum

*INTEGER fields of arbitrary widths (in PER encodings) can be decoded as long as they fit in an i64 (see aper::Decoder::decode_int).

Modules

aper
Tools for encoding and decoding ASN.1 messages of the Aligned PER flavor.

Structs

BitString
A bit string.

Type Definitions

ExtensionMarker
A marker for extendable types.

SequencePreamble
A preamble that describes the properties of the sequence type it preceeds.