Module asn1::aper

A module for encoding and decoding ASN.1 messages of the Aligned PER flavor.

Structs

Constraint
An interval that desribes the limits on some value. To indicate something is unbounded, set min and max to None.

Constraints
A pair of Constraints that describes the constraints on the value (if applicable) and encoded size of a type. A value is considered unconstrained if value and size are both set to None.

Decoder
A bit-wise cursor used to decode aligned PER messagses.

Enums

DecodeError

Constants

UNCONSTRAINED

Traits

APerElement
Trait for Aligned PER encoding/decoding.