Function asn1::aper::encode_int

pub fn encode_int(value: i64, min: Option<i64>, max: Option<i64>) -> Result<Encoding, EncodeError>

Encode an aligned PER integer between min and max.

You can encode the Rust primitive (u)ints: i8, i16, i32, u8, u16, and u32 using their respective to_aper functions. encode_int is useful if you want to encode an integer field that exists somewhere between or beyond the primitive widths.

Examples

For example, a value in [500, 503] can be encoded using two bits in aligned PER, so using u16 would be a waste if bandwidth is precious. The code below demonstrates how to decode such a field.

extern crate asn1;
use asn1::aper::{self, APerElement, Constraint, Constraints, Encoding, encode_int, UNCONSTRAINED};

let x = 501;
println!("{:?}", encode_int(x, Some(500), Some(503).unwrap().bytes()); // Prints [64]Run