Struct asn1::aper::Decoder

pub struct Decoder<'a> { /* fields omitted */ }

A bit-wise cursor used to decode aligned PER messagses.

Methods

impl<'a> Decoder<'a>

fn new(data: &'a [u8]) -> Decoder

Construct a new Decoder with an array of bytes.

fn read(&mut self, n: usize) -> Result<u8, ()>

Read n bits. Where 0 <= n <= 8. See read_to_vec() for larger n. Returns an Err if the read would consume more bits than are available. Else, returns the bits as a u8 with left-padding.

fn read_u8(&mut self) -> Result<u8, ()>

Read a byte.

fn read_to_vec(&mut self, content: &mut Vec<u8>, len: usize) -> Result<(), ()>

Read len bits into content. Returns an Err if the read would consume more bits than are available. Else, the bits as a u8s with left-padding are pushed onto content.

fn decode_length(&mut self) -> Result<usize, DecodeError>

Decode an Aligned PER length determinant

fn decode_int(&mut self, min: Option<i64>, max: Option<i64>) -> Result<i64, DecodeError>

Decode an Aligned PER integer between min and max