Trait asn1::aper::APerElement

pub trait APerElement {
    type Result;

    const TAG: u32;
    const CONSTRAINTS: Constraints;

    fn from_aper(decoder: &mut Decoder, constraints: Constraints) -> Result<Self::Result, DecodeError>;
}

Trait for Aligned PER encoding/decoding.

Examples

Consider a simple ASN.1 Sequence foo made up of a BitString and a 32-bit non-negative integer.

foo ::= SEQUENCE {
    bar BIT STRING(SIZE(4)
    baz INTEGER(0..4294967295)
}Run

The corresponding struct and APerElement implementation are shown below.

#![feature(associated_consts)]
extern crate asn1;
use asn1::BitString;
use asn1::aper::{self, APerElement, Constraint, Constraints, UNCONSTRAINED};

struct foo {
    pub bar: BitString,
    pub baz: u32,
}

impl APerElement for Foo {
   type Result = Self;
   const TAG: u32 = 0xBEEF;
   const CONSTRAINTS: Constraints = UNCONSTRAINED;
   fn from_aper(decoder: &mut aper::Decoder, constraints: Constraints) -> Result<Self::Result, aper::DecodeError> {
       let bar = BitString::from_aper(decoder , Constraints {
           value: None,
           size: Some(Constraint::new(Some(4), Some(4))),
       });

       let mut baz = u32::from_aper(decoder, UNCONSTRAINED);

       if bar.is_err() || baz.is_err() {
           return Err(aper::DecodeError::Dummy);
       }

       Ok(Foo{
           bar: bar.unwrap(),
           baz: baz.unwrap(),
       })
   }
}Run

Now let's consider an enum that corresponds to the ASN.1 Choice type below. (Note the extension marker)

Foo ::= SEQUENCE {
    a BIT STRING(SIZE(4))
}

Bar ::= SEQUENCE {
    a OCTET STRING
}

Baz ::= SEQUENCE {
    a INTEGER(0..255)
    b INTEGER(0..65535)
}

MyMsg ::= CHOICE {
    foo Foo
    bar Bar
    baz Baz
    ...
}Run

The corresponding enum and APerElement implementation would look like this.

#![feature(associated_consts)]
extern crate asn1;
use asn1::BitString;
use asn1::aper::{self, APerElement, Constraint, Constraints, UNCONSTRAINED};

enum MyMsg {
    foo { a: BitString, },
    bar { a: Vec<u8>, },
    baz { a: u8, b: u16, },
}
 
impl APerElement for MyMsg {
    type Result = Self;
    const TAG: u32 = 0xBEEF;
    const CONSTRAINTS: Constraints = UNCONSTRAINED;
    fn from_aper(decoder: &mut aper::Decoder, constraints: Constraints) -> Result<Self::Result, aper::DecodeError> {
        let is_ext = ExtensionMarker::from_aper(decoder, UNCONSTRAINED);
        if is_ext.is_err() {
            return Err(aper::DecodeError::Dummy);
        }
 
        let choice = decoder.decode_int(Some(0), Some(2));
        if choice.is_err() {
            return Err(aper::DecodeError::Dummy);
        }
 
        match c.unwrap() {
            0 => {
                let bs = BitString::from_aper(decoder , Constraints {
                    value: None,
                    size: Some(Constraint::new(None, Some(4))),
                });
                if bs.is_err() {
                    Err(aper::DecodeError::Dummy)
                } else {
                    Ok(MyMsg::foo{ a: bs.unwrap(), })
                }
            },
            1 => {
                let mut v = Vec::<u8>::from_aper(decoder, Constraints {
                    value: None,
                    size: Some(Constraint::new(None, Some(3))),
                });
                if v.is_err() {
                    Err(aper::DecodeError::Dummy)
                } else {
                    Ok(MyMsg::bar{ a: v.unwrap(), })
                }
            },
            2 => {
                let a = u8::from_aper(decoder, UNCONSTRAINED);
                let b = u16::from_aper(decoder, UNCONSTRAINED);
                if a.is_err() || b.is_err() {
                    Err(aper::DecodeError::Dummy)
                } else {
                    Ok(MyMsg::baz{ a: a.unwrap(), b: b.unwrap(), })
                }
            }
            _ => Err(aper::DecodeError::Dummy)
        }
    }
}Run

Associated Types

type Result

The type to be returned by from_aper, usually Self

Associated Constants

const TAG: u32

A tag for determining canonical-PER ordering. Not currently used.

const CONSTRAINTS: Constraints

PER-visible Constraints

Required Methods

fn from_aper(decoder: &mut Decoder, constraints: Constraints) -> Result<Self::Result, DecodeError>

Constructor for the Result type given an aligned PER encoding.