Struct asn1::BitString

pub struct BitString { /* fields omitted */ }

A bit string.

Examples

extern crate asn1;
use asn1::BitString;

let mut b = BitString::with_len(64);
b.set(0, true);
println!("b[0] = {}", b.is_set(0)); // Prints b[0] = trueRun

Methods

impl BitString

fn with_len(n: usize) -> BitString

Consturct a BitString of length n with all values set to 0.

fn with_bytes_and_len(data: &Vec<u8>, n: usize) -> BitString

Consturct a BitString of length n with initial values contained in data.

Examples

extern crate asn1;
use asn1::BitString;
 
let v = vec![0x00 as u8, 0x02 as u8];
let b = BitString::with_bytes_and_len(&v, 15);
println!("b[0] = {}", b.is_set(0)); // Prints b[0] = false
println!("b[14] = {}", b.is_set(14)); // Prints b[14] = trueRun

fn get_num_bits(&self) -> usize

Get the length of a BitString

fn set_num_bits(&mut self, n: usize)

Set the length of a BitString and initialize any new values to 0

fn is_set(&self, i: usize) -> bool

Check if bit i is set.

fn set(&mut self, i: usize, val: bool)

Set bit i to val.